backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "insighthub"
version = "0.1.0"
description = "Business Intelligence Platform for Multi-location Businesses"
readme = "README.md"
authors = [
    {name = "Your Name", email = "you@example.com"}
]
license = {text = "MIT"}
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Office/Business :: Financial",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    # Core
    "Django==4.2.11",
    "djangorestframework==3.14.0",
    "django-cors-headers==4.2.0",
    "django-environ==0.11.2",
    
    # Database
    "psycopg2-binary==2.9.9",
    "django-postgres-extra==2.0.7",
    
    # Data Processing
    "pandas==2.1.4",
    "numpy==1.26.2",
    "scikit-learn==1.3.2",
    "nltk==3.8.1",
    
    # APIs
    "requests==2.31.0",
    "aiohttp==3.9.1",
    "gql==3.4.1",
    
    # Async & Tasks
    "celery==5.3.4",
    "redis==5.0.1",
    "django-celery-beat==2.5.0",
    "django-celery-results==2.5.1",
    
    # API Documentation
    "drf-yasg==1.21.7",
    "django-filter==23.5",
    
    # Authentication
    "djangorestframework-simplejwt==5.3.0",
    "python-jose==3.3.0",
    
    # Monitoring
    "prometheus-client==0.19.0",
    "sentry-sdk==1.40.0",
    
    # Utils
    "python-dateutil==2.8.2",
    "pytz==2023.3",
    "tenacity==8.2.3",
    "cachetools==5.3.2",
]

[project.optional-dependencies]
dev = [
    "black==23.11.0",
    "flake8==6.1.0",
    "mypy==1.7.0",
    "pytest==7.4.3",
    "pytest-django==4.7.0",
    "pytest-asyncio==0.21.1",
    "factory-boy==3.3.0",
    "django-debug-toolbar==4.2.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/insighthub"
"Bug Tracker" = "https://github.com/yourusername/insighthub/issues"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true